refactor(rust-core): split lib.rs into canonical/context/oracle/result (internal only, API stable)
